Unify KV cache and logits/probs storage on bf16.
Replace fp16 (device half) storage with bf16 (device ushort +
arena_load/arena_store) across the KV cache and logits/probs pipelines so
the whole pipeline uses bf16 uniformly. Previously these buffers stored
IEEE fp16 with a bf16-round pre-pass; now they store bf16 bits directly,
matching the rest of the arena.

Metal (12 kernels): attention, qk_rope_kv, pack_encoder_kv (KV cache);
softcap_half, half_scale_fp16, sc_probs, sc_prob_cols, sc_softembed,
sample_rowstats, sample_apply, scatter_logits_rows, logit_rowstats
(logits/probs). device half -> device ushort, float(x[i]) -> arena_load,
half(v) -> arena_store.

Rust: step_kv.rs f32_to_half_bits -> bf16::f32_to_bf16_bits; readback
f16_bits_to_f32 -> bf16_bits_to_f32 (step_kv.rs and step_kernel.rs
read_half_buffer_f32/half_buffer_stats/count_non_finite_half/
read_logit_f32); removed now-dead f16_bits_to_f32 from step_kernel.rs.

Test fixtures + CPU oracles: flipped fp16 round-trips to bf16 round-trips
so GPU/CPU parity holds under the new storage format.

Out of scope (left as-is): GEMM threadgroup half tiles (MMA input
precision - fp16's 10 mantissa bits beneficial) and the separate f32
lm_head GPU path.

1.71 -> 1.70 s/step; all in-scope tier-1 tests (attention, qk_rope_kv,
sc_*, sample_*, softcap_half, logit_rowstats, step_kv, step_kernel) and
fusion_matches_unfused_forward_logits pass. Pre-existing gemm_q8 and
sample::stopper failures unchanged.

+++ shaders/kernels/scatter_logits_rows.metal
@@ -1,10 +1,12 @@
 #include <metal_stdlib>
 using namespace metal;
 
+#include "arena.metal"
+
 /// Scatter compact `[num_active, vocab]` logits rows into full `[canvas, vocab]` buffer.
 kernel void scatter_logits_rows(
-    device const half *compact [[buffer(0)]],
-    device half *logits [[buffer(1)]],
+    device const ushort *compact [[buffer(0)]],
+    device ushort *logits [[buffer(1)]],
     device const uint *row_map [[buffer(2)]],
     constant uint2 &dims [[buffer(3)]],
     uint2 gid [[thread_position_in_grid]]
@@ -17,5 +19,6 @@ kernel void scatter_logits_rows(
         return;
     }
     uint dst_row = row_map[bi];
+    // bf16 bits copy verbatim (no rounding needed — pure bit move).
     logits[(ulong)dst_row * vocab + col] = compact[(ulong)bi * vocab + col];
 }
